requirements.txt
discord.py==0.16.6
PyYAML==3.12
requests==2.13.0
redis==2.10.5
